refactor proposal version replay evidence router
