+++ src/app/router_registry.py
@@ -387,6 +387,9 @@ from app.routers.proposal_report_requests import router as proposal_report_reque
 from app.routers.proposal_risk_approval import router as proposal_risk_approval_router
 from app.routers.proposal_version_async import router as proposal_version_async_router
 from app.routers.proposal_version_commands import router as proposal_version_commands_router
+from app.routers.proposal_version_replay_evidence import (
+    router as proposal_version_replay_evidence_router,
+)
 from app.routers.proposal_versions import router as proposal_versions_router
 from app.routers.proposal_workflow import router as proposal_workflow_router
 from app.routers.proposal_workflow_decisions import router as proposal_workflow_decisions_router
@@ -508,6 +511,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(proposal_versions_router)
     app.include_router(proposal_version_async_router)
     app.include_router(proposal_version_commands_router)
+    app.include_router(proposal_version_replay_evidence_router)
     app.include_router(proposal_workflow_router)
     app.include_router(proposal_workflow_decisions_router)
     app.include_router(proposal_risk_approval_router)
